Decompose users from ads-app to users-app
