+++ ads/models.py
@@ -22,33 +22,6 @@ class Location(models.Model):
         }
 
 
-class User(models.Model):
-    first_name = models.CharField(max_length=50)
-    last_name = models.CharField(max_length=50)
-    username = models.CharField(max_length=50)
-    password = models.CharField(max_length=50)
-    role = models.CharField(max_length=20)
-    age = models.PositiveSmallIntegerField()
-    location = models.ForeignKey(Location, null=True, on_delete=models.SET_NULL)
-
-    class Meta:
-        verbose_name = "Пользователь"
-        verbose_name_plural = "Пользователи"
-
-    def __str__(self):
-        return f"{self.id}: {self.first_name} {self.last_name}"
-
-    def get_dict(self):
-        return {
-            'id': self.id,
-            'first_name': self.first_name,
-            'last_name': self.last_name,
-            'username': self.username,
-            'role': self.role,
-            'age': self.age,
-            'location': self.location.get_dict(),
-        }
-
 
 class Category(models.Model):
     name = models.CharField(max_length=100, unique=True)
@@ -73,7 +46,7 @@ class Advertisement(models.Model):
         verbose_name_plural = "Объявления"
 
     name = models.CharField(max_length=100)
-    author = models.ForeignKey(User, on_delete=models.CASCADE)
+    author = models.ForeignKey('users.User', on_delete=models.CASCADE)
     price = models.PositiveIntegerField()
     description = models.CharField(max_length=2000)
     image = models.ImageField(upload_to='images/')
