+++ project/urls.py
@@ -20,5 +20,6 @@ import ads.views
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', ads.views.index),
-    path('', include('ads.urls'))
+    path('', include('ads.urls')),
+    path('', include('users.urls')),
 ]
